Refactor shared ReductionCore code into a base algorithm class
