+++ Framework/PythonInterface/plugins/algorithms/WorkflowAlgorithms/SANS/SANSReductionCoreEventSlice.py
@@ -11,17 +11,15 @@ for which data must be event sliced. These steps are: slicing, adjustment, conve
 
 from __future__ import (absolute_import, division, print_function)
 
-from mantid.api import (DistributedDataProcessorAlgorithm, MatrixWorkspaceProperty, AlgorithmFactory, PropertyMode,
-                        IEventWorkspace, Progress)
+from mantid.api import (MatrixWorkspaceProperty, AlgorithmFactory, PropertyMode,
+                        Progress)
 from mantid.kernel import (Direction, PropertyManagerProperty, StringListValidator)
-from sans.algorithm_detail.mask_workspace import mask_bins
-from sans.common.constants import EMPTY_NAME
 from sans.common.enums import (DetectorType, DataType)
-from sans.common.general_functions import append_to_sans_file_tag, create_child_algorithm
-from sans.state.state_base import create_deserialized_sans_state_from_property_manager
+
+from SANSReductionCoreBase import SANSReductionCoreBase
 
 
-class SANSReductionCoreEventSlice(DistributedDataProcessorAlgorithm):
+class SANSReductionCoreEventSlice(SANSReductionCoreBase):
     def category(self):
         return 'SANS\\Reduction'
 
@@ -80,24 +78,8 @@ class SANSReductionCoreEventSlice(DistributedDataProcessorAlgorithm):
         # ----------
         # OUTPUT
         # ----------
-        self.declareProperty(MatrixWorkspaceProperty("OutputWorkspace", '', direction=Direction.Output),
-                             doc='The output workspace.')
-
-        self.declareProperty(MatrixWorkspaceProperty('SumOfCounts', '', optional=PropertyMode.Optional,
-                                                     direction=Direction.Output),
-                             doc='The sum of the counts of the output workspace.')
-
-        self.declareProperty(MatrixWorkspaceProperty('SumOfNormFactors', '', optional=PropertyMode.Optional,
-                                                     direction=Direction.Output),
-                             doc='The sum of the counts of the output workspace.')
-
-        self.declareProperty(MatrixWorkspaceProperty('CalculatedTransmissionWorkspace', '', optional=PropertyMode.Optional,
-                                                     direction=Direction.Output),
-                             doc='The calculated transmission workspace')
-
-        self.declareProperty(MatrixWorkspaceProperty('UnfittedTransmissionWorkspace', '', optional=PropertyMode.Optional,
-                                                     direction=Direction.Output),
-                             doc='The unfitted transmission workspace')
+        # SANSReductionCoreEventSlice has the same outputs as SANSReductionCore
+        self._pyinit_output()
 
     def PyExec(self):
         # Get the input
@@ -176,139 +158,6 @@ class SANSReductionCoreEventSlice(DistributedDataProcessorAlgorithm):
         self.setProperty("CalculatedTransmissionWorkspace", calculated_transmission_workspace)
         self.setProperty("UnfittedTransmissionWorkspace", unfitted_transmission_workspace)
 
-    def _slice(self, state_serialized, workspace, monitor_workspace, data_type_as_string):
-        slice_name = "SANSSliceEvent"
-        slice_options = {"SANSState": state_serialized,
-                         "InputWorkspace": workspace,
-                         "InputWorkspaceMonitor": monitor_workspace,
-                         "OutputWorkspace": EMPTY_NAME,
-                         "OutputWorkspaceMonitor": "dummy2",
-                         "DataType": data_type_as_string}
-        slice_alg = create_child_algorithm(self, slice_name, **slice_options)
-        slice_alg.execute()
-
-        workspace = slice_alg.getProperty("OutputWorkspace").value
-        monitor_workspace = slice_alg.getProperty("OutputWorkspaceMonitor").value
-        slice_event_factor = slice_alg.getProperty("SliceEventFactor").value
-        return workspace, monitor_workspace, slice_event_factor
-
-    def _adjustment(self, state_serialized, workspace, monitor_workspace, component_as_string, data_type):
-        transmission_workspace = self._get_transmission_workspace()
-        direct_workspace = self._get_direct_workspace()
-
-        adjustment_name = "SANSCreateAdjustmentWorkspaces"
-        adjustment_options = {"SANSState": state_serialized,
-                              "Component": component_as_string,
-                              "DataType": data_type,
-                              "MonitorWorkspace": monitor_workspace,
-                              "SampleData": workspace,
-                              "OutputWorkspaceWavelengthAdjustment": EMPTY_NAME,
-                              "OutputWorkspacePixelAdjustment": EMPTY_NAME,
-                              "OutputWorkspaceWavelengthAndPixelAdjustment": EMPTY_NAME}
-        if transmission_workspace:
-            transmission_workspace = self._move(state_serialized, transmission_workspace, component_as_string,
-                                                is_transmission=True)
-            adjustment_options.update({"TransmissionWorkspace": transmission_workspace})
-
-        if direct_workspace:
-            direct_workspace = self._move(state_serialized, direct_workspace, component_as_string, is_transmission=True)
-            adjustment_options.update({"DirectWorkspace": direct_workspace})
-
-        adjustment_alg = create_child_algorithm(self, adjustment_name, **adjustment_options)
-        adjustment_alg.execute()
-
-        wavelength_adjustment = adjustment_alg.getProperty("OutputWorkspaceWavelengthAdjustment").value
-        pixel_adjustment = adjustment_alg.getProperty("OutputWorkspacePixelAdjustment").value
-        wavelength_and_pixel_adjustment = adjustment_alg.getProperty(
-                                           "OutputWorkspaceWavelengthAndPixelAdjustment").value
-        calculated_transmission_workspace = adjustment_alg.getProperty("CalculatedTransmissionWorkspace").value
-        unfitted_transmission_workspace = adjustment_alg.getProperty("UnfittedTransmissionWorkspace").value
-        return wavelength_adjustment, pixel_adjustment, wavelength_and_pixel_adjustment, \
-            calculated_transmission_workspace, unfitted_transmission_workspace
-
-    def _move(self, state_serialized, workspace, component, is_transmission=False):
-        # First we set the workspace to zero, since it might have been moved around by the user in the ADS
-        # Second we use the initial move to bring the workspace into the correct position
-        move_name = "SANSMove"
-        move_options = {"SANSState": state_serialized,
-                        "Workspace": workspace,
-                        "MoveType": "SetToZero",
-                        "Component": ""}
-        move_alg = create_child_algorithm(self, move_name, **move_options)
-        move_alg.execute()
-        workspace = move_alg.getProperty("Workspace").value
-
-        # Do the initial move
-        move_alg.setProperty("MoveType", "InitialMove")
-        move_alg.setProperty("Component", component)
-        move_alg.setProperty("Workspace", workspace)
-        move_alg.setProperty("IsTransmissionWorkspace", is_transmission)
-        move_alg.execute()
-        return move_alg.getProperty("Workspace").value
-
-    def _get_transmission_workspace(self):
-        transmission_workspace = self.getProperty("TransmissionWorkspace").value
-        return self._get_cloned_workspace(transmission_workspace) if transmission_workspace else None
-
-    def _get_direct_workspace(self):
-        direct_workspace = self.getProperty("DirectWorkspace").value
-        return self._get_cloned_workspace(direct_workspace) if direct_workspace else None
-
-    def _convert_to_histogram(self, workspace):
-        if isinstance(workspace, IEventWorkspace):
-            convert_name = "RebinToWorkspace"
-            convert_options = {"WorkspaceToRebin": workspace,
-                               "WorkspaceToMatch": workspace,
-                               "OutputWorkspace": "OutputWorkspace",
-                               "PreserveEvents": False}
-            convert_alg = create_child_algorithm(self, convert_name, **convert_options)
-            convert_alg.execute()
-            workspace = convert_alg.getProperty("OutputWorkspace").value
-            append_to_sans_file_tag(workspace, "_histogram")
-
-        return workspace
-
-    def _copy_bin_masks(self, workspace, dummy_workspace):
-        mask_options = {"InputWorkspace": workspace,
-                        "MaskedWorkspace": dummy_workspace,
-                        "OutputWorkspace": EMPTY_NAME}
-        mask_alg = create_child_algorithm(self, "MaskBinsFromWorkspace", **mask_options)
-        mask_alg.execute()
-        return mask_alg.getProperty("OutputWorkspace").value
-
-    def _convert_to_q(self, state_serialized, workspace, wavelength_adjustment_workspace, pixel_adjustment_workspace,
-                      wavelength_and_pixel_adjustment_workspace):
-        """
-        A conversion to momentum transfer is performed in this step.
-
-        The conversion can be either to the modulus of Q in which case the output is a 1D workspace, or it can
-        be a 2D reduction where the y axis is Qy, ie it is a numeric axis.
-        @param state: a SANSState object
-        @param workspace: the workspace to convert to momentum transfer.
-        @param wavelength_adjustment_workspace: the wavelength adjustment workspace.
-        @param pixel_adjustment_workspace: the pixel adjustment workspace.
-        @param wavelength_and_pixel_adjustment_workspace: the wavelength and pixel adjustment workspace.
-        @return: a reduced workspace
-        """
-        convert_name = "SANSConvertToQ"
-        convert_options = {"InputWorkspace": workspace,
-                           "OutputWorkspace": EMPTY_NAME,
-                           "SANSState": state_serialized,
-                           "OutputParts": True}
-        if wavelength_adjustment_workspace:
-            convert_options.update({"InputWorkspaceWavelengthAdjustment": wavelength_adjustment_workspace})
-        if pixel_adjustment_workspace:
-            convert_options.update({"InputWorkspacePixelAdjustment": pixel_adjustment_workspace})
-        if wavelength_and_pixel_adjustment_workspace:
-            convert_options.update({"InputWorkspaceWavelengthAndPixelAdjustment":
-                                    wavelength_and_pixel_adjustment_workspace})
-        convert_alg = create_child_algorithm(self, convert_name, **convert_options)
-        convert_alg.execute()
-        data_workspace = convert_alg.getProperty("OutputWorkspace").value
-        sum_of_counts = convert_alg.getProperty("SumOfCounts").value
-        sum_of_norms = convert_alg.getProperty("SumOfNormFactors").value
-        return data_workspace, sum_of_counts, sum_of_norms
-
     def validateInputs(self):
         errors = dict()
         # Check that the input can be converted into the right state object
@@ -318,24 +167,6 @@ class SANSReductionCoreEventSlice(DistributedDataProcessorAlgorithm):
         except ValueError as err:
             errors.update({"SANSSingleReductionEventSlice": str(err)})
         return errors
-
-    def _get_state(self):
-        state_property_manager = self.getProperty("SANSState").value
-        state = create_deserialized_sans_state_from_property_manager(state_property_manager)
-        state.property_manager = state_property_manager
-        return state
-
-    def _get_monitor_workspace(self):
-        monitor_workspace = self.getProperty("ScatterMonitorWorkspace").value
-        return self._get_cloned_workspace(monitor_workspace)
-
-    def _get_cloned_workspace(self, workspace):
-        clone_name = "CloneWorkspace"
-        clone_options = {"InputWorkspace": workspace,
-                         "OutputWorkspace": EMPTY_NAME}
-        clone_alg = create_child_algorithm(self, clone_name, **clone_options)
-        clone_alg.execute()
-        return clone_alg.getProperty("OutputWorkspace").value
 
     def _get_progress(self):
         return Progress(self, start=0.0, end=1.0, nreports=5)
